% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marcus-package.R
\docType{package}
\name{A.marcus-package}
\alias{A.marcus-package}
\alias{marcus}
\title{Marcus Institute for Aging Internal Package (package)}
\description{
Marcus Institute for Aging Internal Package (package)
}
\keyword{package}
