% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_redcap.R
\name{read_redcap}
\alias{read_redcap}
\title{Read data from IFAR REDCap}
\usage{
read_redcap(token, ids = NULL, fields = NULL, events = NULL)
}
\arguments{
\item{token}{Your REDCap token for specific project}

\item{ids}{A vector of selected participants ID to be pulled. If \code{NULL}
(default), all records will be pulled.}

\item{fields}{A vector of selected fields to be pulled. Default is \code{NULL}.}

\item{events}{A vector of selected study events. Default is \code{NULL}.}
}
\description{
Wrapper of REDCapR::redcap_read_oneshot
}
