% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_personal_r_packages.R
\name{remove_personal_r_packages}
\alias{remove_personal_r_packages}
\title{Remove All personal R packages}
\usage{
remove_personal_r_packages()
}
\description{
This RStudio Addin cleans up the installations of all personal
R packages.
}
